filter chip 분리
